provision-testing-environment: patch pkg_resources.py bug.
The version of python3-pkg-resources shipped on Precise has a bug which
causes spurious output which breaks some of our tests:

Issue #341 in distribute (< 0.6.33).
See: https://bitbucket.org/tarek/distribute/issue/341
See also comments in this merge request:
https://code.launchpad.net/~sylvain-pineau/checkbox/xlsx_exporter/+merge/166032

This applies a patch to fix this bug "on the fly", on the testing
container, to avoid having to work around that in our code. This is
attempted only if we're using LXC (for some reason it doesn't happen on
Vagrant), and if it fails we only report it and continue on our way. The
assumption is that, if it failed, it's a non-Precise release where the
fix is unlikely to be needed. This avoids introducing release-detection
logic.