checkbox-support: resurrect the submission parser
This patch adds a verbatim copy of the submission parser (except for tests
where the import parts where altered for checkbox_support). The code
doesn't pass test as it imports lots of bits and pieces from the checkbox
python package (now checkbox-legacy project on Launchpad).

Those will be replaced by code from chcekbox-support over the next several
patches, so that changes from the old submission parser can be easily
identified.

This patch also brings in the dependency on libxml2, again

Fixes: https://bugs.launchpad.net/checkbox-support/+bug/1312808

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>