checkbox-support: copy TestRun class from hexr
This patch adds a bit of hexr code to the submission parser. This is done
for two reasons. First the TestRun is required to obtain the actual
'parsing' of the submission in a way that hexr and c3 expects to.

With that done, hexr can actually import checkbox_support and use
the parser directly. Better yet, we can now own both ends of
the submission pipe so the internal format (the xml file) can be changed
and improved (we can drop the whole XML madness) as ad the end of the day
the certification website expects a nice and tidy JSON object with
a list of JSON objects and XML is not needed to transport those.

This will allow us to drop the libxml2 dependency (along with a number
of security issues that it keeps having), simplify the whole process
by not having to do any eventual-no-op XML transformations and we won't
have to rely on complex XML validation.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>