checkbox-support:parsers:udevadm: use type("") instead of str
This is a litlte hack for python2.7 compatibility. The udevadm parser
can be invoked from 'plainbox dev parse udevadm' where it gets a pre-loaded
string object (python3). In python2.7 we may pass an unicode object instead
and the code would assume that's a stream object (which it clearly isnt)
so this hack allows us to avoid having to spell either basestr or unicode

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>