test-in-lxc: More reliable container destruction.
lxc-destroy -f gets stuck sometimes when using a memory-backed,
ephemeral VM. To avoid this, we first try to lxc-stop the container.
Ephemeral containers disappear when stopped. We wait for the container's
state to indicate it has stopped, and then, if still present (indicating
a non-memory-backed container) we destroy it, which should work all the
time, to free up disk resources.