plainbox:ctrl: inject PLAINBOX_PROVIDER_UNITS into execution environment
This patch adds PLAINBOX_PROVIDER_UNITS to the list of environment
variables that are added to the execution environment of all job commands.
The variable can be used to look at all the units (for example, as a
reliable way to look at job definitions and later on, whitelists).
The variable is only added if the provider.units_dir directory is set.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>