plainbox:secure:providers: discover jobs by unit.unit, not by class
Since the JobDefinition class is just an implementation detail we should
not use it for 'is this unit a job definition unit' test. Instead, we
can rely on the new unit property to check if a job is a job. This limits
the quirkiness of the current arrangement where jobs don't need to define
the unit type to work.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>