checkbox-support:parsers:submission: Be smarter about EFI parsing.
As seen in
https://bugs.launchpad.net/plainbox-provider-checkbox/+bug/1334131 the
submission parser blindly invokes the UEFI/BIOS subparser on any
attachment that contains the string 'UEFI'. This converts the simple "X
in Y" matching into regular expression matching, then adds a negative
lookahead to the string for EFI attachments, to ensure we don't parse
the "uefirtvariable" one by mistake.

Adding this kind of specificity in the submission parser is somewhat
kludgy, but the context parser selection is already very hardcoded and
based on odd heuristics, so in a sense it's consistent.