plainbox:unit: add Unit.get_translated_record_value()
This patch adds a high-level version of a translated-record-attribute
lookup function. It is meant to replace all of the manually implemented
tr_foo() class functions (except for the special tr_unit()).

Unlike all of the current implementations it also understands unit
parameters so it should be have well when used with templates.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>