checkbox-touch: import and use pyotherside 1.2
This patch adds a minimalistic Python component that just logs a few
things to stdout about its existence and the versions involved.

It doesn't work out of the box though, on a touch device you need
lib/arm-linux-gnueabihf/ pre-populated with additional things that
we haven't yet automated. It works well on the desktop providing you have
pyotherside installed from the archive.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>