checkbox-touch:components: add PythonObjectHandle
This QML component complements the recently added python APIs.
The component wraps a single handle and pyotherside reference and
exposes one public function, invoke(). That function allows one to
asynchronously call any method on the wrapped python object.

Internally the _unref() function is called when the QML component is
being destroyed and the handleReady() signal is sent when the component
becomes usable (typically all user initialization code would be in
the handler of that signal)

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>