checkbox-touch:components: add WelcomePage
This welcome page implements the current design (we should make the
design public for easier access) of the touch UI. The welcome text
is still as stub (waiting for design update).

The button action is exposed via the startTestingTriggered() signal The
text of the welcome message is exposed as the welcomeText property
alias.

The QML code was heavily improved by Nekhelesh Ramananthan. I'd like to
also thank Michael Zanetti for his suggestions.

Improved-by: Nekhelesh Ramananthan <krnekhelesh@gmail.com>
Improved-by: Michael Zanetti <michael.zanetti@canonical.com>
Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>