plainbox:unit: add UnitWithId
This patch adds a subclass of Unit (that is not registered as an actual unit
one can use) which adds an id field along with the typical semantics. It can be
used by all units that need to have an identifier but want to avoid code
duplication.

The JobDefinition class currently *doesn't* use this base class as it has some
custom partial_id construction logic but this can be fixed later on.

Signed-off-by: Zygmunt Krynicki <zkrynicki@gmail.com>