support: Generate changelogs for all components being released.
A logs/ directory will contain logs for each packaging branch and for
trunk.

Logs are built from the latest stable tag in each branch (if doing a
release, this is the latest tag *before* the one being released).

Logs should be checked manually and can be integrated in a release
announcement.

The provided bzr log formatter will be used if it can be set up;
otherwise, we fall back to standard bzr --short format, which is
suboptimal but should work.