checkbox-support:parsers:udevadm: Add heuristics for QEMU/KVM virtio disks.
These heuristics properly categorize these devices (category DISK). By
their nature they're very bare and have few identifying features, but the
driver and bus are quite distinctive, and the rest of the attributes used
should also be present and reliable.

The rest of the tests in the test suite were not modified, to ensure
the new heuristics don't miscategorize any existing or previously-seen
systems and/or devices.