plainbox:unit:job: silence warnings from JobDefinition.estimated_duration
This patch silences the annoying warning that was printed if the
'estimated_duration' field was defined but incorrect. That case is
now checked by the new job validator.

The error case is still "valid" as in, no exception is being raised
as not everything is currently using the new validator.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>