plainbox:unit:file: add FileUnit
This patch adds FileUnit. A virtual unit (one that is not typically expected to
be created by anyone explicitly) for describing files that belong to a
provider. The unit has two essential fields: path and role. Path is simply
pointing to an actual file. The role describes the way that particular file is
expected to be used.

Currently the following roles are defined:
 - unit definition file (unit-source)
 - legacy whitelist (whitelist)
 - architecture independent executable (script)
 - architecture dependent executable (binary)
 - data file (data)
 - translation catalog (i18n)
 - management script (manage.py)
 - license & copyright (legal)
 - documentation (docs)

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>