plainbox:session:state: remove logging in SessionState.on_{job,unit}_{added,removed}
This patch removes logging in the SessionState.on_{job,unit}_{added,removed}
signals. The rationale is that it is better to "see" the same signals in the
session device context where, at least, we know which context we are talking
about.

For all the code using the new context APIs it also removes the annoying
feeling that certain events are logged twice.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>