plainbox:session:manager: teach session manager about device contexts
This large patch is changing the SessionManager class from a dumb tuple
(state, storage) into a less-dumb tuple (device_context_list, storage).

The primary responsibility is still the same, that is, to combine the
in-memory state (SessionState) with the in-filesystem state
(SessionStorage) but this patch begins the emerging new responsibility
of offering the only needed API for everything involving a "session".
Subsequent patches will add a way to run jobs and thus directly remove
the need to force applications to manage a JobRunner.

The most basic change that this patch is doing is to replace a single
SessionState with a list of SessionDeviceContext objects. As you know,
SessionDeviceContext wraps a SessionState (for now) so this is
equivalent to just having a list of states. Since each context also
holds a reference to a Device object (currently all devices are None)
this will allow us to keep the SessionManager as the only required API
for working with more complicated sessions (remote testing and
multi-node testing).

Looking deeper there are a lot of lesser changes:

    The initializer is different, as it now takes a device_context_list
    argument instead of a state object. There is a new @classmethod
    create() method for constructing an empty manager.

    There are two new methods that allow applications to add or remove a
    device context dynamically. Note that because the on-disk format is
    not yet ready for multi-node testing (or devices in general) using
    more than one device context is currently prohibited.

    There are also new signals whenever a device context is added or
    removed that are tied to the two methods mentioned earlier.

    For convenience there is a simple API to add a device context
    corresponding to the current device (which will not have to chance
    after the devices patches become available and land in trunk). All
    applications that don't care about remote testing or multi-node
    testing can use that method and not have to deal with any
    lower-level APIs.

    For similar convenience there is a way to access the "default"
    device context that simply maps to the first available context, if
    any.

    The old property that exposed the session state object
    (SessionManager.state) is now implemented on top of the
    default_device_context

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>