checkbox-touch: rewrite and integrate with plainbox
This mammoth patch changes checkbox-touch's main.qml, all of the py/
code and tweaks the remaining QML files to work together for real.

NOTE: you will need to have latest version of plainbox (not 0.12)
around. This will be fixed by issuing a separate mini-release just for
testing the touch version.

There are countless changes here:
 - The QML side of the application now uses request() which has a way to
   return errors as well. This might be a candidate to merge with
   invoke() since we never use invoke apart from inside request()
 - The python side of the application has a bit different API, object
   identifiers are passed around in a number of places.
 - Logging is significantly better, both on the QML/javascript side and
   on the python side. It's still a little bit too verbose for
   production mode (DEBUG!) but it's readable enough to follow what the
   application is doing.
 - Application start-up is easier to understand. The interactions
   between Python, CheckboxStack and CheckboxTouchApplication are mostly
   gone now. Only Python and CheckboxTouchApplication currently remain
   but I will probably eradicate C-T-A entirely and just use one Python
   component for everything as it is just a layer of confusing
   indirection.
 - Life cycle of python objects is easier to follow with the new log
   calls in __new__ and __del__.
 - Application is parametrizable by editing the .desktop file (kind of
   like a poor man's launcher). Just edit the test plan identifier and
   you're done. Current desktop file uses the 2013.c.c.plainbox::stub
   test plan for a pretty good initial test coverage.
 - The initial welcome screen is changed to let us load everything.
   There's a new activity indicator and the application only allows you
   to proceed after everything is ready.
 - The entries of sys.path that are problematic on a apparmour-sandboxed
   Ubuntu environment are removed.
 - Logging is disabled since plainbox opens some file-based logging
   handlers that don't work on confinement.
 - Tests that fail to execute something (EPERM) are marked as failed

Some squashed commits that I wanted to preserve are added below:

This patch adds the request() method, which builds upon the lower-level
invoke() method. Unlike invoke, request gets two callback functions, one
for successful returns and another one for failed returns.

This can simplify error handling and lessen the need to handle
asynchronous exception events.

The downside is that now all return values must be in a form of a
dictionary with one of two possible "schemas". For successful calls the
callee must return:

response = {
   'code': 200,
   'result': ...
}

For errors the callee must return:

response = {
   'code': 500,
   'error': ...
}

checkbox-touch:py: add the @view decorator

This decorator implements a simple convention for all view functions that
must return a dictionary that is expected by the request() javascript
function, that is, one that has the 'code' and either 'error' or 'result'

checkbox-touch:py: print a full traceback when @view fails

This patch makes anything decorated with a @view decorator print a full
traceback in the event of an uncaught exception. This is much better than
the short one-line error str() that is printed by pyotherside.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>