plainbox:session:legacy: fix regression in _commit_{open,clean}()
This patch fixes a regression in the legacy compatibility API caused
by "64e2a40 plainbox:session:manager: teach session manager about
device contexts" where the SessionManager lost the _state attribute
and re-wired the SessionManager.state property to return the session
state of the default device context.

This change was harmless, except that SessionStateLegacyAPICompatImpl's
_commit_open() and _commit_clean() were relying on being able to overwrite
the _state attribute with a different object. Now the actual "valuable"
state was stored in a now-unused attribute and the state that was
saved to disk was a vanilla, empty one.

The fix is pretty easy, as now SessionManager has the .create_with_state()
constructor we can just use that and dispense with the hackery. The same
trick was used by _commit_open() and _commit_clean() so two identical
changes are made. In addition to that, since we are no longer "hacking"
anything, the long, descriptive comment was removed since the description
of the create_with_state() method is sufficient to explain the intended
semantics.

Fixes: https://bugs.launchpad.net/plainbox/+bug/1375158

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>