checkbox-ng: Honor submission/submit_to_hexr launcher setting.
Added the required handling in newcli.py, and added a config variable called
submission/submit_to_hexr so this can be used in the config file. As is
customary, the launcher will override the config file setting. I tested both
a launcher and adding the setting in plainbox.conf.

Fixes https://bugs.launchpad.net/checkbox-ng/+bug/1375854