plainbox:commands:checkbox,checkbox:commands:sru: add the --test-plan option
This patch adds a new option to the 'checkbox' base command that is used
by commands such as 'plainbox run', 'checkbox sru' or 'plainbox dev
analyze' to expose the new --test-plan option.

This option acts similarly to the existing --whitelist option but has
three important differences. First of all, it supports the new test
plans with all the bells and whistles. Second of all, it looks up the
test plan by ID, not by a file path. Lastly, it can only be provided
once.

The last limitation can be lifted at a later time but for now it is
unclear how multiple test plans should co-exist in one session with
regards to things like category overrides.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>