plainbox:session:resume: always retain desired jobs
This patch changes the ResumeDiscardQualifier to retain jobs mentioned
in both the 'jobs' structure (that maps job IDs to checksums) as well as
in the list of desired jobs ('desired_job_list').

This seems like a no-op but it is actually relevant when the 'jobs' data
structure stops to store jobs that don't have any results yet (as is the
case in the 4th suspend format). In such cases jobs will be always
mentioned in the desired_job_list (by their identifier) but they might
not be yet mentioned in the jobs' mapping as they may not have a result
yet.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>