plainbox:session:suspend: add 4th suspend format
This patch adds the 4th version of the peek helper.  The fourth format
is not saving hollow results. This means that a session with a 1000 jobs
gets saved to almost nothing as all the results are initially hollow.
More than than, jobs with only hollow results are *not* mentioned in the
id -> checksum map. This allows us to safely upgrade jobs that don't
effectively participate in a session without loosing the ability to
resume that session later.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>