plainbox:commands:parse: split into cmd+inv
This patch splits the 'parse.py' module into two new modules, cmd_parse
and inv_parse. During plainbox startup only the cmd_parse module needs
to be imported and it has minimal requirements so very little overall
code needs to be loaded.

When the 'plainbox dev parse' command is executed then the inv_parse
module is imported, bringing in all of the additional modules it
requires.

In addition, all call sites that used to import the old module are
updated to use one of the new modules instead.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>