plainbox:commands:session: split into cmd+inv
This patch splits the 'session.py' module into two new modules,
cmd_session and inv_session. During plainbox startup only the
cmd_session module needs to be imported and it has minimal requirements
so very little overall code needs to be loaded.

When the session command is actually executed then the inv_session
module is imported, bringing in all of the additional modules it
requires.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>