plainbox:commands:special: split into cmd+inv
This patch splits the 'special.py' module into two new modules,
cmd_special and inv_special. During plainbox startup only the
cmd_special module needs to be imported and it has minimal requirements
so very little overall code needs to be loaded.

When the 'plainbox dev special' command is executed then the inv_special
module is imported, bringing in all of the additional modules it
requires.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>