plainbox:commands:startprovider: split into cmd+inv
This patch splits the 'startprovider.py' module into two new modules,
cmd_startprovider and inv_startprovider. During plainbox startup only
the cmd_startprovider module needs to be imported and it has minimal
requirements so very little overall code needs to be loaded.

When the 'plainbox self-test' command is executed then the
inv_startprovider module is imported, bringing in all of the additional
modules it requires.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>